refactor: extract approval lifecycle
